remove more whacky proj stuff and fix some 64bit warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/branches/dll@4913 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/vc/dirent.c
@@ -79,7 +79,7 @@ struct dirent *	readdir(DIR *dirp) {
     memset(&d,'\0', sizeof(struct dirent));
 
     strcpy(d.d_name,dirp->findFileData.cFileName);
-    d.d_namlen = strlen(d.d_name);
+    d.d_namlen = (unsigned int)strlen(d.d_name);
 
     return &d;
 }
